Move config into scripts
