--- config/jest/cssTransform.js
@@ -1,10 +1,0 @@
-'use strict'
-
-// This is a custom Jest transformer turning style imports into empty objects.
-// http://facebook.github.io/jest/docs/en/webpack.html
-
-module.exports = {
-  process() {
-    return 'module.exports = {};'
-  },
-}
