+++ scripts/config/cachedIconData.json
