+++ scripts/config/paths.js
