+++ scripts/start.js
@@ -10,7 +10,7 @@ process.on('unhandledRejection', (err) => {
 })
 
 // Ensure environment variables are read.
-require('../config/env')
+require('./config/env')
 
 const fs = require('fs')
 const chalk = require('chalk')
@@ -26,9 +26,9 @@ const {
   prepareUrls,
 } = require('react-dev-utils/WebpackDevServerUtils')
 const openBrowser = require('react-dev-utils/openBrowser')
-const paths = require('../config/paths')
-const config = require('../config/webpack.config.dev')
-const createDevServerConfig = require('../config/webpackDevServer.config')
+const paths = require('./config/paths')
+const config = require('./config/webpack.config.dev')
+const createDevServerConfig = require('./config/webpackDevServer.config')
 
 const useYarn = fs.existsSync(paths.yarnLockFile)
 const isInteractive = process.stdout.isTTY
